start taking zsh apart

+++ zsh/.config/zsh/.zshrc
@@ -1,9 +1,9 @@
+export DOTFILES="$HOME/.dotfiles"
 export XDG_CACHE_HOME="$HOME/.cache"
 export XDG_CONFIG_HOME="$HOME/.config"
 export XDG_STATE_HOME="$HOME/.local/state"
 export XDG_DATA_HOME="$HOME/.local/share"
 export HISTFILE="$XDG_STATE_HOME"/zsh/history
-export CARGO_HOME="$XDG_DATA_HOME"/cargo
 export CUDA_CACHE_PATH="$XDG_CACHE_HOME"/nv
 export GNUPGHOME="$XDG_DATA_HOME"/gnupg
 export GOPATH="$XDG_DATA_HOME"/go
@@ -25,10 +25,17 @@ export STARSHIP_CONFIG=~/.config/starship/starship.toml
 export DOTNET_ROOT=/opt/dotnet
 export BROWSER="/usr/bin/firefox"
 export GPGKEY=
-export CARGO_TARGET_DIR="$XDG_DATA_HOME/cargo_target"
-export ZELLIJ_AUTO_ATTACH=0
 
-export PATH=$HOME/bin:/usr/local/bin:$PATH:$HOME/.local/share/nvim/mason/bin/:$HOME/.local/share/cargo/bin/:$XDG_DATA_HOME/zplug/bin/:
+export PATH=$HOME/bin:/usr/local/bin:$PATH:$HOME/.local/share/nvim/mason/bin/:$XDG_DATA_HOME/zplug/bin/:
+
+# Cargo
+if command -v cargo >/dev/null 2>&1; then
+  export CARGO_HOME="$XDG_DATA_HOME"/cargo
+  export CARGO_TARGET_DIR="$XDG_DATA_HOME/cargo_target"
+  export PATH=$PATH:$HOME/.local/share/cargo/bin/
+
+  alias cargi='f() { cargo install "\$1" && echo "\$1" >> ~/.dotfiles/rust/cargo_packages; unset -f f; }; f'
+fi
 
 # fuzzy finder
 export FZF_DEFAULT_OPTS='-i --height=50%'
@@ -55,13 +62,26 @@ zplug "zsh-users/zsh-history-substring-search", defer:3
 # Then, source plugins and add commands to $PATH
 zplug load
 
+# Should be called before compinit
+zmodload zsh/complist
+source "$XDG_CONFIG_HOME/zsh/completion.zsh"
+
+# Use hjlk in menu selection (during completion)
+# Doesn't work well with interactive mode
+bindkey -M menuselect 'h' vi-backward-char
+bindkey -M menuselect 'k' vi-up-line-or-history
+bindkey -M menuselect 'j' vi-down-line-or-history
+bindkey -M menuselect 'l' vi-forward-char
+
+bindkey -M menuselect '^xg' clear-screen
+bindkey -M menuselect '^xi' vi-insert                     # Insert
+bindkey -M menuselect '^xh' accept-and-hold               # Hold
+bindkey -M menuselect '^xn' accept-and-infer-next-history # Next
+bindkey -M menuselect '^xu' undo                          # Undo
+
 ## ENVIRONEMENT VARIALES
-export MonolithRoot=/home/$USER/Development/devinite/
 export SWAYSOCK=/run/user/$(id -u)/sway-ipc.$(id -u).$(pgrep -x sway).sock
 export EDITOR=nvim
-
-# Compilation flags
-# export ARCHFLAGS="-arch x86_64"
 
 # custom functions
 function yy() {
@@ -73,55 +93,69 @@ function yy() {
   rm -f -- "$tmp"
 }
 
-# Set personal aliases, overriding those provided by oh-my-zsh libs,
-# plugins, and themes. Aliases can be placed here, though oh-my-zsh
-# users are encouraged to define aliases within the ZSH_CUSTOM folder.
-# For a full list of active aliases, run `alias`.
-#
-# Example aliases
-# alias zshconfig="mate ~/.zshrc"
-# alias ohmyzsh="mate ~/.oh-my-zsh"
-alias l='exa --icons --git --no-user'
-alias ll='exa -l --icons --git -a'
-alias lt='exa --tree --level=2 --long --icons --git'
+if command -v exa >/dev/null 2>&1; then
+  alias l='exa --icons --git --no-user'
+  alias ll='exa -l --icons --git -a'
+  alias lt='exa --tree --level=2 --long --icons --git'
+else
+  alias l='ls'
+  alias ll='ls -la'
+fi
+
+if command -v fdfind >/dev/null 2>&1; then
+  alias fd='fdfind'
+fi
+
+if command -v bat >/dev/null 2>&1; then
+  alias cat='bat --style=plain'
+fi
+
+if command -v fzf >/dev/null 2>&1; then
+  bindkey -s '^v' "fzf --preview 'bat --color=always {}' \n"
+  alias txkill="tmux ls | fzf | awk '{print $1;}' | xargs -n1 tmux kill-session -t"
+
+  eval "$(fzf --zsh)"
+fi
+
+# Swap <Esc> and <Capslock> for better vim experience on foreign keyboards
 alias swapesc='/usr/bin/setxkbmap -option "caps:swapescape"'
-alias kdev='kubectl config use-context $(kubectl config get-contexts -o name | grep core-dev)'
-alias ktest='kubectl config use-context $(kubectl config get-contexts -o name | grep core-test)'
-alias kprod='kubectl config use-context $(kubectl config get-contexts -o name | grep core-prod)'
-alias k9='k9s -n dg-sales'
 alias v="nvim"
-alias ze="zellij"
 alias wget=wget --hsts-file="$XDG_DATA_HOME/wget-hsts"
 alias oscpvpn="sudo -b openvpn --config ~/projects/offsec/universal.ovpn "
-alias sales="cd $HOME/Development/Dg.Sales/ && zj attach sales || zj -s sales"
-alias config="cd $HOME/.dotfiles/neovim/.config/nvim/ && zj attach config || zj -s config"
 alias cd..="cd .."
 alias cd2="cd ../.."
 alias cd3="cd ../../.."
 alias cd4="cd ../../../.."
-alias txkill="tmux ls | fzf | awk '{print $1;}' | xargs -n1 tmux kill-session -t"
-alias cat='bat --style=plain'
-alias fd='fdfind'
 alias pbcopy='xsel --input --clipboard'
 alias pbpaste='xsel --output --clipboard'
 alias ip='ip -c'
-alias cargi='f() { cargo install "\$1" && echo "\$1" >> ~/.dotfiles/rust/cargo_packages; unset -f f; }; f'
-bindkey -s '^f' "zellij-sessionizer /home/$USER \n"
-bindkey -s '^v' "fzf --preview 'bat --color=always {}' \n"
 
-[ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"  # This loads nvm
-[ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"  # This loads nvm bash_completion
+[ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"                   # This loads nvm
+[ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion" # This loads nvm bash_completion
 
+export WORKRC="$HOME/.config/zsh/.workrc.sh"
+# Work specific configs
+if [[ -f $WORKRC ]]; then
+  source $WORKRC
+fi
+
+if command -v zellij >/dev/null 2>&1; then
+  source "$XDG_CONFIG_HOME/zellij/zellij.zsh"
+fi
 
 # Created by `pipx` on 2023-04-18 06:15:55
 export PATH="$PATH:/home/$USER/.local/bin"
 export PATH="$PATH:/opt/mssql-tools18/bin"
 
-eval "$(zoxide init zsh)"
+if command -v zoxide >/dev/null 2>&1; then
+  eval "$(zoxide init zsh)"
+fi
+
+if command -v starship >/dev/null 2>&1; then
+  eval "$(starship init zsh)"
+fi
 eval "$(starship init zsh)"
 
 #THIS MUST BE AT THE END OF THE FILE FOR SDKMAN TO WORK!!!
 export SDKMAN_DIR="$HOME/.sdkman"
 [[ -s "$HOME/.sdkman/bin/sdkman-init.sh" ]] && source "$HOME/.sdkman/bin/sdkman-init.sh"
-
-eval "$(fzf --zsh)"
